hero: extract BlockEditor from Entities so a block editor can live in a modal (phase 4)
Preview-first editor groundwork: pull the per-block editor rendering (the
type switch over CommonEntity / FeaturedProductBase) and its media/product
handlers out of Entities into a standalone BlockEditor({index, entity,
featuredProducts}) that reads the form via useFormContext. Entities renders it
inline exactly as before — no behaviour change — but the same component can now
be dropped into the click-to-edit modal (next commit). Handlers are scoped to
the block's index; form paths stay positional, product cache keys by uid.

tsc 27 -> 27, yarn build passes.
