+++ src/components/managers/hero/components/entities.tsx
@@ -13,15 +13,13 @@ import {
   sortableKeyboardCoordinates,
   verticalListSortingStrategy,
 } from '@dnd-kit/sortable';
-import { common_MediaFull } from 'api/proto-http/admin';
 import { heroTypes } from 'constants/constants';
 import { FC, useCallback, useEffect, useRef, useState } from 'react';
 import { useFormContext, useWatch } from 'react-hook-form';
 import { Button } from 'ui/components/button';
 import Text from 'ui/components/text';
 import { EntitiesProps } from '../utility/interface';
-import { CommonEntity } from './common-entity';
-import { FeaturedProductBase } from './featured-prduct-base';
+import { BlockEditor } from './block-editor';
 import { HeroSchema } from './schema';
 import { SortableEntity } from './sortable-entity';
 
@@ -33,7 +31,6 @@ export const Entities: FC<EntitiesProps> = ({
   onDeletedIndicesChange,
 }) => {
   const {
-    setValue,
     control,
     formState: { errors },
   } = useFormContext<HeroSchema>();
@@ -105,76 +102,6 @@ export const Entities: FC<EntitiesProps> = ({
     });
   }, [entities]);
 
-  const handleSaveMedia = useCallback(
-    (
-      selectedMedia: common_MediaFull[],
-      entityIndex: number,
-      type: 'main' | 'single' | 'doubleLeft' | 'doubleRight',
-      orientation: 'Portrait' | 'Landscape',
-    ) => {
-      if (!selectedMedia.length) return;
-
-      const media = selectedMedia[0];
-      const thumbnailUrl = media.media?.thumbnail?.mediaUrl || '';
-
-      let idPath: string;
-      let urlPath: string;
-
-      if (type === 'doubleLeft' || type === 'doubleRight') {
-        const side = type === 'doubleLeft' ? 'left' : 'right';
-        idPath = `entities.${entityIndex}.double.${side}.media${orientation}Id`;
-        urlPath = `entities.${entityIndex}.double.${side}.media${orientation}Url`;
-      } else {
-        const entityType = type === 'main' ? 'main' : 'single';
-        idPath = `entities.${entityIndex}.${entityType}.media${orientation}Id`;
-        urlPath = `entities.${entityIndex}.${entityType}.media${orientation}Url`;
-      }
-
-      setValue(idPath as any, media.id, { shouldDirty: true, shouldTouch: true });
-      setValue(urlPath as any, thumbnailUrl, { shouldDirty: true, shouldTouch: true });
-    },
-    [setValue],
-  );
-
-  const handleClearMedia = useCallback(
-    (
-      entityIndex: number,
-      type: 'main' | 'single' | 'doubleLeft' | 'doubleRight',
-      orientation: 'Portrait' | 'Landscape',
-    ) => {
-      let idPath: string;
-      let urlPath: string;
-
-      if (type === 'doubleLeft' || type === 'doubleRight') {
-        const side = type === 'doubleLeft' ? 'left' : 'right';
-        idPath = `entities.${entityIndex}.double.${side}.media${orientation}Id`;
-        urlPath = `entities.${entityIndex}.double.${side}.media${orientation}Url`;
-      } else {
-        const entityType = type === 'main' ? 'main' : 'single';
-        idPath = `entities.${entityIndex}.${entityType}.media${orientation}Id`;
-        urlPath = `entities.${entityIndex}.${entityType}.media${orientation}Url`;
-      }
-
-      setValue(idPath as any, undefined, { shouldDirty: true, shouldTouch: true });
-      setValue(urlPath as any, '', { shouldDirty: true, shouldTouch: true });
-    },
-    [setValue],
-  );
-
-  const handleSaveProductSelection = useCallback(
-    (newProducts: any[], index: number, uid: string) => {
-      const productIds = newProducts
-        .map((product) => product.id)
-        .filter((id): id is number => id !== undefined);
-
-      // form path is positional; display cache is keyed by the stable uid
-      setValue(`entities.${index}.featuredProducts.productIds` as any, productIds);
-      featuredProducts.saveSelection(newProducts, uid);
-      featuredProducts.closeSelection();
-    },
-    [setValue, featuredProducts],
-  );
-
   const handleRemoveEntity = useCallback((uid: string) => {
     setDeletedIndices((prev) => new Set(prev).add(uid));
   }, []);
@@ -201,116 +128,6 @@ export const Entities: FC<EntitiesProps> = ({
     if (fromIndex === toIndex) return;
 
     arrayHelpers.move(fromIndex, toIndex);
-  };
-
-  const renderEntity = (entity: HeroSchema['entities'][number], index: number) => {
-    const uid = (entity as any)._uid as string;
-    switch (entity.type) {
-      case 'HERO_TYPE_MAIN':
-        return (
-          <CommonEntity
-            title='main add'
-            prefix={`entities.${index}.main`}
-            landscapeLink={entity.main?.mediaLandscapeUrl || ''}
-            portraitLink={entity.main?.mediaPortraitUrl || ''}
-            aspectRatio={{
-              Portrait: ['9:16'],
-              Landscape: ['2:1'],
-            }}
-            onSaveMedia={(media: common_MediaFull[], orientation: 'Portrait' | 'Landscape') =>
-              handleSaveMedia(media, index, 'main', orientation)
-            }
-            onClearMedia={(orientation) => handleClearMedia(index, 'main', orientation)}
-          />
-        );
-
-      case 'HERO_TYPE_SINGLE':
-        return (
-          <CommonEntity
-            title='single add'
-            prefix={`entities.${index}.single`}
-            landscapeLink={entity.single?.mediaLandscapeUrl || ''}
-            portraitLink={entity.single?.mediaPortraitUrl || ''}
-            aspectRatio={{
-              Portrait: ['9:16'],
-              Landscape: ['2:1'],
-            }}
-            onSaveMedia={(media: common_MediaFull[], orientation: 'Portrait' | 'Landscape') =>
-              handleSaveMedia(media, index, 'single', orientation)
-            }
-            onClearMedia={(orientation) => handleClearMedia(index, 'single', orientation)}
-          />
-        );
-
-      case 'HERO_TYPE_DOUBLE':
-        return (
-          <div className='flex flex-col gap-4'>
-            <div className='lg:px-2.5 p-2.5'>
-              <Text className='text-xl font-bold leading-none' variant='uppercase'>
-                double add
-              </Text>
-            </div>
-            <CommonEntity
-              title='left add'
-              prefix={`entities.${index}.double.left`}
-              landscapeLink={entity.double?.left?.mediaLandscapeUrl || ''}
-              portraitLink={entity.double?.left?.mediaPortraitUrl || ''}
-              aspectRatio={['1:1']}
-              isDoubleAd={true}
-              onSaveMedia={(media: common_MediaFull[], orientation: 'Portrait' | 'Landscape') =>
-                handleSaveMedia(media, index, 'doubleLeft', orientation)
-              }
-              onClearMedia={(orientation) => handleClearMedia(index, 'doubleLeft', orientation)}
-            />
-            <CommonEntity
-              title='right add'
-              prefix={`entities.${index}.double.right`}
-              landscapeLink={entity.double?.right?.mediaLandscapeUrl || ''}
-              portraitLink={entity.double?.right?.mediaPortraitUrl || ''}
-              aspectRatio={['1:1']}
-              isDoubleAd
-              onSaveMedia={(media: common_MediaFull[], orientation: 'Portrait' | 'Landscape') =>
-                handleSaveMedia(media, index, 'doubleRight', orientation)
-              }
-              onClearMedia={(orientation) => handleClearMedia(index, 'doubleRight', orientation)}
-            />
-          </div>
-        );
-
-      case 'HERO_TYPE_FEATURED_PRODUCTS':
-        return (
-          <FeaturedProductBase
-            index={index}
-            uid={uid}
-            entity={entity}
-            product={featuredProducts.products}
-            currentEntityUid={featuredProducts.currentUid}
-            isModalOpen={featuredProducts.isOpen}
-            showProductPicker={true}
-            title='featured products'
-            prefix='featuredProducts'
-            handleOpenProductSelection={featuredProducts.openSelection}
-            handleCloseModal={featuredProducts.closeSelection}
-            handleSaveNewSelection={handleSaveProductSelection}
-            handleProductsReorder={featuredProducts.reorderProducts}
-          />
-        );
-
-      case 'HERO_TYPE_FEATURED_PRODUCTS_TAG':
-        return (
-          <FeaturedProductBase
-            index={index}
-            uid={uid}
-            entity={entity}
-            product={{}}
-            title='featured products tag'
-            prefix='featuredProductsTag'
-          />
-        );
-
-      default:
-        return null;
-    }
   };
 
   return (
@@ -416,7 +233,15 @@ export const Entities: FC<EntitiesProps> = ({
                         </div>
                       </div>
 
-                      {!isCollapsed && <div>{renderEntity(entity, index)}</div>}
+                      {!isCollapsed && (
+                        <div>
+                          <BlockEditor
+                            index={index}
+                            entity={entity}
+                            featuredProducts={featuredProducts}
+                          />
+                        </div>
+                      )}
                     </div>
                   );
                 }}
